...pvmanager: merge with 1.0-SNAPSHOT.

+++ core/plugins/org.csstudio.utility.pvmanager/src/org/epics/pvmanager/jca/DataUtils.java
@@ -71,7 +71,7 @@ class DataUtils {
      * @return a new pvmanager timestamp
      */
     static org.epics.pvmanager.util.TimeStamp fromEpics(gov.aps.jca.dbr.TimeStamp epicsTimeStamp) {
-        return org.epics.pvmanager.util.TimeStamp.time(epicsTimeStamp.secPastEpoch() + TS_EPOCH_SEC_PAST_1970, epicsTimeStamp.nsec());
+        return org.epics.pvmanager.util.TimeStamp.time(epicsTimeStamp.secPastEpoch() + TS_EPOCH_SEC_PAST_1970, (int) epicsTimeStamp.nsec());
     }
 
     /**
